EY-3209 Sikre at OMS-søknad kun journalføres én gang
